ci: publish to Cocoapods when release PR is merged
A release tag must be created before the Cocoapods publishing happens
since the XCTFramework file needs to be uploaded to the release tag,
and the generated link used in the .podspec file to make the binary
available on cocoapods.

In order to automate the process, knowing that the tag is created
first, this repository will rely on a `release/<version>` branch being
merged into `main` to trigger the automation.